feat: add helper function and refactor readme
