+++ priceloop_api/utils/default_configuration.py
@@ -1,7 +1,7 @@
 import priceloop_api
 from typing import Callable
 
-from priceloop_api.util.auth_state import AuthState
+from priceloop_api.utils.auth_state import AuthState
 
 default_host_name = 'alpha.priceloop.ai'
